requirements.txt
Django==3.2.7
djangorestframework
drf-yasg
gunicorn
polls
markdown
django-filter
psycopg2
django-environ
django-cors-headers
PyJWT
django-rest-auth 
django-allauth
djangorestframework-jwt